+++ apps/token/app/Cargo.toml
@@ -1,12 +1,8 @@
 [package]
 edition = "2021"
 name = "token"
-version = "0.1.0"
 publish = false
-
-[[bin]]
-name = "app"
-path = "src/main.rs"
+version = "0.1.0"
 
 [dependencies]
 anyhow = { workspace = true }
@@ -15,7 +11,7 @@ essential-app-utils = { workspace = true }
 essential-hash = { workspace = true }
 essential-rest-client = { workspace = true }
 essential-sign = { workspace = true }
-# essential-signer = { workspace = true }
+essential-signer = { workspace = true }
 essential-types = { workspace = true }
 essential-wallet = { workspace = true }
 hex = { workspace = true }
@@ -26,5 +22,8 @@ tokio = { workspace = true }
 
 [dev-dependencies]
 essential-app-utils = { workspace = true, features = ["test-utils"] }
-
-# essential-wallet = { workspace = true, features = ["test-utils"] }
+essential-wallet = { workspace = true, features = ["test-utils"] }
+essential-builder = { workspace = true, features = ["tracing"] }
+essential-builder-db.workspace = true
+essential-node = { workspace = true, features = ["tracing"] }
+tracing-subscriber.workspace = true
