+++ apps/token/app/src/mint.rs
@@ -1,3 +1,6 @@
+//! # Mint
+//! Contains functionality for minting new tokens in the token contract.
+
 use essential_app_utils::inputs::Encode;
 use essential_sign::secp256k1::ecdsa::RecoverableSignature;
 use essential_types::{
@@ -8,34 +11,52 @@ use essential_types::{
 
 use crate::{balance, nonce, Query};
 
+/// Represents the initial data required for minting tokens.
 pub struct Init {
+    /// The hashed key of the account.
     pub hashed_key: [Word; 4],
+    /// The amount of tokens to mint.
     pub amount: Word,
+    /// The number of decimals of the token.
     pub decimals: Word,
+    /// The current nonce of the account.
     pub nonce: Query,
 }
 
+/// Represents the data to be signed for a mint solution.
 pub struct ToSign {
+    /// The hashed key of the account.
     pub hashed_key: [Word; 4],
+    /// The amount of tokens to mint.
     pub amount: Word,
+    /// The number of decimals of the token.
     pub decimals: Word,
+    /// The new nonce of the account.
     pub new_nonce: Word,
 }
 
+/// Contains all necessary information to build a mint solution.
 pub struct BuildSolution {
+    /// The new nonce of the account.
     pub new_nonce: Word,
+    /// The current balance of the account.
     pub current_balance: Query,
+    /// The hashed key of the account.
     pub hashed_key: [Word; 4],
+    /// The amount of tokens to mint.
     pub amount: Word,
+    /// The number of decimals of the token.
     pub decimals: Word,
+    /// The signature over the data.
     pub signature: RecoverableSignature,
+    /// The name of the token.
     pub token_name: String,
+    /// The symbol of the token.
     pub token_symbol: String,
 }
 
-pub struct Submit(pub Solution);
-
 impl ToSign {
+    /// Converts the ToSign struct to a vector of Words for signing.
     pub fn to_words(&self) -> Vec<Word> {
         vec![
             self.hashed_key[0],
@@ -49,6 +70,7 @@ impl ToSign {
     }
 }
 
+/// Prepares the data to be signed for a mint transaction.
 pub fn data_to_sign(account: Init) -> anyhow::Result<ToSign> {
     let Init {
         hashed_key,
@@ -65,6 +87,7 @@ pub fn data_to_sign(account: Init) -> anyhow::Result<ToSign> {
     })
 }
 
+/// Builds a mint solution based on the provided data.
 pub fn build_solution(build: BuildSolution) -> anyhow::Result<Solution> {
     let BuildSolution {
         new_nonce,
@@ -103,10 +126,12 @@ pub fn build_solution(build: BuildSolution) -> anyhow::Result<Solution> {
     })
 }
 
+/// Increments the nonce by 1.
 fn increment_nonce(nonce: Word) -> Word {
     nonce + 1
 }
 
+/// Calculates the new balance after minting tokens.
 fn calculate_new_balance(balance: Word, amount: Word) -> anyhow::Result<Word> {
     balance
         .checked_add(amount)
