+++ apps/token/app/src/transfer.rs
@@ -1,3 +1,6 @@
+//! # Transfer
+//! Contains functionality for transferring tokens between accounts in the token contract.
+
 use essential_app_utils::inputs::Encode;
 use essential_sign::secp256k1::ecdsa::RecoverableSignature;
 use essential_types::{
@@ -7,33 +10,50 @@ use essential_types::{
 
 use crate::{balance, nonce, Query};
 
+/// Represents the initial data required for transferring tokens.
 pub struct Init {
+    /// The hashed key of the sender.
     pub hashed_from_key: [Word; 4],
+    /// The hashed key of the recipient.
     pub hashed_to_key: [Word; 4],
+    /// The amount of tokens to transfer.
     pub amount: Word,
+    /// The current nonce of the sender.
     pub nonce: Query,
 }
 
+/// Represents the data to be signed for a transfer solution.
 pub struct ToSign {
+    /// The hashed key of the sender.
     pub hashed_from_key: [Word; 4],
+    /// The hashed key of the recipient.
     pub hashed_to_key: [Word; 4],
+    /// The amount of tokens to transfer.
     pub amount: Word,
+    /// The new nonce of the sender.
     pub new_nonce: Word,
 }
 
+/// Contains all necessary information to build a transfer solution.
 pub struct BuildSolution {
+    /// The hashed key of the sender.
     pub hashed_from_key: [Word; 4],
+    /// The hashed key of the recipient.
     pub hashed_to_key: [Word; 4],
+    /// The new nonce of the sender.
     pub new_nonce: Word,
+    /// The amount of tokens to transfer.
     pub amount: Word,
+    /// The current balance of the sender.
     pub current_from_balance: Query,
+    /// The current balance of the recipient.
     pub current_to_balance: Query,
+    /// The signature over the data.
     pub signature: RecoverableSignature,
 }
 
-pub struct Submit(pub Solution);
-
 impl ToSign {
+    /// Converts the ToSign struct to a vector of Words for signing.
     pub fn to_words(&self) -> Vec<Word> {
         vec![
             self.hashed_from_key[0],
@@ -50,6 +70,7 @@ impl ToSign {
     }
 }
 
+/// Prepares the data to be signed for a transfer solution.
 pub fn data_to_sign(account: Init) -> anyhow::Result<ToSign> {
     let Init {
         hashed_from_key,
@@ -66,6 +87,7 @@ pub fn data_to_sign(account: Init) -> anyhow::Result<ToSign> {
     })
 }
 
+/// Builds a transfer solution based on the provided data.
 pub fn build_solution(build: BuildSolution) -> anyhow::Result<Solution> {
     let BuildSolution {
         hashed_from_key,
@@ -104,16 +126,19 @@ pub fn build_solution(build: BuildSolution) -> anyhow::Result<Solution> {
     })
 }
 
+/// Increments the nonce by 1.
 fn increment_nonce(nonce: Word) -> Word {
     nonce + 1
 }
 
+/// Calculates the new balance for the sender after transferring tokens.
 fn calculate_from_balance(from_balance: Word, amount: Word) -> anyhow::Result<Word> {
     from_balance
         .checked_sub(amount)
         .ok_or(anyhow::anyhow!("Insufficient balance"))
 }
 
+/// Calculates the new balance for the recipient after receiving tokens.
 fn calculate_to_balance(to_balance: Word, amount: Word) -> anyhow::Result<Word> {
     to_balance
         .checked_add(amount)
