--- apps/token/app2/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-edition = "2021"
-name = "token"
-publish = false
-version = "0.1.0"
-
-[dependencies]
-anyhow = { workspace = true }
-clap = { workspace = true }
-essential-app-utils = { workspace = true }
-essential-hash = { workspace = true }
-essential-rest-client = { workspace = true }
-essential-sign = { workspace = true }
-essential-signer = { workspace = true }
-essential-types = { workspace = true }
-essential-wallet = { workspace = true }
-hex = { workspace = true }
-pint-abi = { workspace = true }
-rpassword = { workspace = true }
-serde_json = { workspace = true }
-tokio = { workspace = true }
-
-[dev-dependencies]
-essential-app-utils = { workspace = true, features = ["test-utils"] }
-essential-wallet = { workspace = true, features = ["test-utils"] }
-essential-builder = { workspace = true, features = ["tracing"] }
-essential-builder-db.workspace = true
-essential-node = { workspace = true, features = ["tracing"] }
-tracing-subscriber.workspace = true
